Remove code duplication
