+++ airbyte-integrations/connectors/source-salesforce/unit_tests/integration/test_bulk_stream.py
@@ -3,22 +3,17 @@
 import json
 import urllib.parse
 from datetime import datetime, timedelta, timezone
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Any, Dict, Optional
 from unittest import TestCase
 
 import freezegun
-from airbyte_cdk.sources.source import TState
-from airbyte_cdk.test.catalog_builder import CatalogBuilder
-from airbyte_cdk.test.entrypoint_wrapper import EntrypointOutput, read
 from airbyte_cdk.test.mock_http import HttpMocker, HttpRequest, HttpResponse
-from airbyte_cdk.test.mock_http.request import ANY_QUERY_PARAMS
-from airbyte_cdk.test.state_builder import StateBuilder
-from airbyte_protocol.models import ConfiguredAirbyteCatalog, SyncMode
+from airbyte_protocol.models import SyncMode
 from config_builder import ConfigBuilder
-from integration.utils import given_stream
+from integration.utils import given_authentication, given_stream, read
 from salesforce_describe_response_builder import SalesforceDescribeResponseBuilder
-from source_salesforce import SourceSalesforce
 from source_salesforce.streams import LOOKBACK_SECONDS
+
 
 _A_FIELD_NAME = "a_field"
 _ACCESS_TOKEN = "an_access_token"
@@ -34,37 +29,6 @@ _REFRESH_TOKEN = "a_refresh_token"
 _STREAM_NAME = "a_stream_name"
 
 _BASE_URL = f"{_INSTANCE_URL}/services/data/{_API_VERSION}"
-
-
-def _catalog(sync_mode: SyncMode) -> ConfiguredAirbyteCatalog:
-    return CatalogBuilder().with_stream(_STREAM_NAME, sync_mode).build()
-
-
-def _source(catalog: ConfiguredAirbyteCatalog, config: Dict[str, Any], state: Optional[TState]) -> SourceSalesforce:
-    return SourceSalesforce(catalog, config, state)
-
-
-def _read(
-    sync_mode: SyncMode,
-    config_builder: Optional[ConfigBuilder] = None,
-    state_builder: Optional[StateBuilder] = None,
-    expecting_exception: bool = False
-) -> EntrypointOutput:
-    catalog = _catalog(sync_mode)
-    config = config_builder.build() if config_builder else ConfigBuilder().build()
-    state = state_builder.build() if state_builder else StateBuilder().build()
-    return read(_source(catalog, config, state), config, catalog, state, expecting_exception)
-
-
-def _given_authentication(http_mocker: HttpMocker, client_id: str, client_secret: str, refresh_token: str) -> None:
-    http_mocker.post(
-        HttpRequest(
-            "https://login.salesforce.com/services/oauth2/token",
-            query_params=ANY_QUERY_PARAMS,
-            body=f"grant_type=refresh_token&client_id={client_id}&client_secret={client_secret}&refresh_token={refresh_token}"
-        ),
-        HttpResponse(json.dumps({"access_token": _ACCESS_TOKEN, "instance_url": _INSTANCE_URL})),
-    )
 
 
 def _create_field(name: str, _type: Optional[str] = None) -> Dict[str, Any]:
@@ -93,7 +57,7 @@ class FullRefreshTest(TestCase):
 
     @HttpMocker()
     def test_when_read_then_create_job_and_extract_records_from_result(self, http_mocker: HttpMocker) -> None:
-        _given_authentication(http_mocker, _CLIENT_ID, _CLIENT_SECRET, _REFRESH_TOKEN)
+        given_authentication(http_mocker, _CLIENT_ID, _CLIENT_SECRET, _REFRESH_TOKEN, _INSTANCE_URL)
         given_stream(http_mocker, _BASE_URL, _STREAM_NAME, SalesforceDescribeResponseBuilder().field(_A_FIELD_NAME))
         http_mocker.post(
             HttpRequest(f"{_BASE_URL}/jobs/query", body=json.dumps({"operation": "queryAll", "query": "SELECT a_field FROM a_stream_name", "contentType": "CSV", "columnDelimiter": "COMMA", "lineEnding": "LF"})),
@@ -108,6 +72,6 @@ class FullRefreshTest(TestCase):
             HttpResponse(f"{_A_FIELD_NAME}\nfield_value"),
         )
 
-        output = _read(SyncMode.full_refresh, self._config)
+        output = read(_STREAM_NAME, SyncMode.full_refresh, self._config)
 
         assert len(output.records) == 1
