+++ airbyte-integrations/connectors/source-salesforce/unit_tests/integration/test_rest_stream.py
@@ -3,26 +3,21 @@
 import json
 import urllib.parse
 from datetime import datetime, timedelta, timezone
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Any, Dict, Optional
 from unittest import TestCase
 
 import freezegun
-from airbyte_cdk.sources.source import TState
-from airbyte_cdk.test.catalog_builder import CatalogBuilder
-from airbyte_cdk.test.entrypoint_wrapper import EntrypointOutput, read
 from airbyte_cdk.test.mock_http import HttpMocker, HttpRequest, HttpResponse
-from airbyte_cdk.test.mock_http.request import ANY_QUERY_PARAMS
+
 from airbyte_cdk.test.state_builder import StateBuilder
-from airbyte_protocol.models import ConfiguredAirbyteCatalog, SyncMode
+from airbyte_protocol.models import SyncMode
 from config_builder import ConfigBuilder
-from integration.utils import given_stream
+from integration.utils import given_authentication, given_stream, read
 from salesforce_describe_response_builder import SalesforceDescribeResponseBuilder
-from source_salesforce import SourceSalesforce
 from source_salesforce.api import UNSUPPORTED_BULK_API_SALESFORCE_OBJECTS
 from source_salesforce.streams import LOOKBACK_SECONDS
 
 _A_FIELD_NAME = "a_field"
-_ACCESS_TOKEN = "an_access_token"
 _API_VERSION = "v57.0"
 _CLIENT_ID = "a_client_id"
 _CLIENT_SECRET = "a_client_secret"
@@ -33,37 +28,6 @@ _LOOKBACK_WINDOW = timedelta(seconds=LOOKBACK_SECONDS)
 _NOW = datetime.now(timezone.utc)
 _REFRESH_TOKEN = "a_refresh_token"
 _STREAM_NAME = UNSUPPORTED_BULK_API_SALESFORCE_OBJECTS[0]
-
-
-def _catalog(sync_mode: SyncMode) -> ConfiguredAirbyteCatalog:
-    return CatalogBuilder().with_stream(_STREAM_NAME, sync_mode).build()
-
-
-def _source(catalog: ConfiguredAirbyteCatalog, config: Dict[str, Any], state: Optional[TState]) -> SourceSalesforce:
-    return SourceSalesforce(catalog, config, state)
-
-
-def _read(
-    sync_mode: SyncMode,
-    config_builder: Optional[ConfigBuilder] = None,
-    state_builder: Optional[StateBuilder] = None,
-    expecting_exception: bool = False
-) -> EntrypointOutput:
-    catalog = _catalog(sync_mode)
-    config = config_builder.build() if config_builder else ConfigBuilder().build()
-    state = state_builder.build() if state_builder else StateBuilder().build()
-    return read(_source(catalog, config, state), config, catalog, state, expecting_exception)
-
-
-def _given_authentication(http_mocker: HttpMocker, client_id: str, client_secret: str, refresh_token: str) -> None:
-    http_mocker.post(
-        HttpRequest(
-            "https://login.salesforce.com/services/oauth2/token",
-            query_params=ANY_QUERY_PARAMS,
-            body=f"grant_type=refresh_token&client_id={client_id}&client_secret={client_secret}&refresh_token={refresh_token}"
-        ),
-        HttpResponse(json.dumps({"access_token": _ACCESS_TOKEN, "instance_url": _INSTANCE_URL})),
-    )
 
 
 def _create_field(name: str, _type: Optional[str] = None) -> Dict[str, Any]:
@@ -92,7 +56,7 @@ class FullRefreshTest(TestCase):
 
     @HttpMocker()
     def test_given_error_on_fetch_chunk_when_read_then_retry(self, http_mocker: HttpMocker) -> None:
-        _given_authentication(http_mocker, _CLIENT_ID, _CLIENT_SECRET, _REFRESH_TOKEN)
+        given_authentication(http_mocker, _CLIENT_ID, _CLIENT_SECRET, _REFRESH_TOKEN, _INSTANCE_URL)
         given_stream(http_mocker, _BASE_URL, _STREAM_NAME, SalesforceDescribeResponseBuilder().field(_A_FIELD_NAME))
         http_mocker.get(
             HttpRequest(f"{_INSTANCE_URL}/services/data/{_API_VERSION}/queryAll?q=SELECT+{_A_FIELD_NAME}+FROM+{_STREAM_NAME}+"),
@@ -102,7 +66,7 @@ class FullRefreshTest(TestCase):
             ]
         )
 
-        output = _read(SyncMode.full_refresh, self._config)
+        output = read(_STREAM_NAME, SyncMode.full_refresh, self._config)
 
         assert len(output.records) == 1
 
@@ -115,7 +79,7 @@ class IncrementalTest(TestCase):
         self._http_mocker = HttpMocker()
         self._http_mocker.__enter__()
 
-        _given_authentication(self._http_mocker, _CLIENT_ID, _CLIENT_SECRET, _REFRESH_TOKEN)
+        given_authentication(self._http_mocker, _CLIENT_ID, _CLIENT_SECRET, _REFRESH_TOKEN, _INSTANCE_URL)
         given_stream(self._http_mocker, _BASE_URL, _STREAM_NAME, SalesforceDescribeResponseBuilder().field(_A_FIELD_NAME).field(_CURSOR_FIELD, "datetime"))
 
     def tearDown(self) -> None:
@@ -132,7 +96,7 @@ class IncrementalTest(TestCase):
             HttpResponse(json.dumps({"records": [{"a_field": "a_value"}]})),
         )
 
-        _read(SyncMode.incremental, self._config, StateBuilder().with_stream_state(_STREAM_NAME, {}))
+        read(_STREAM_NAME, SyncMode.incremental, self._config, StateBuilder().with_stream_state(_STREAM_NAME, {}))
 
         # then HTTP requests are performed
 
@@ -145,7 +109,7 @@ class IncrementalTest(TestCase):
             HttpResponse(json.dumps({"records": [{"a_field": "a_value"}]})),
         )
 
-        output = _read(SyncMode.incremental, self._config, StateBuilder().with_stream_state(_STREAM_NAME, {_CURSOR_FIELD: cursor_value.isoformat(timespec="milliseconds")}))
+        output = read(_STREAM_NAME, SyncMode.incremental, self._config, StateBuilder().with_stream_state(_STREAM_NAME, {_CURSOR_FIELD: cursor_value.isoformat(timespec="milliseconds")}))
 
         assert output.most_recent_state.stream_state.dict() == {"state_type": "date-range", "slices": [{"start": _to_partitioned_datetime(start), "end": _to_partitioned_datetime(_NOW)}]}
 
@@ -174,7 +138,7 @@ class IncrementalTest(TestCase):
             HttpResponse(json.dumps({"records": [{"a_field": "a_value"}]})),
         )
 
-        output = _read(SyncMode.incremental, self._config, state)
+        output = read(_STREAM_NAME, SyncMode.incremental, self._config, state)
 
         # the start is granular to the second hence why we have `000` in terms of milliseconds
         assert output.most_recent_state.stream_state.dict() == {"state_type": "date-range", "slices": [{"start": _to_partitioned_datetime(start), "end": _to_partitioned_datetime(_NOW)}]}
